removed duplicated pieces from vscode-json-languageservice from YLS
* Remove the copied JSON parser and its direct parser tests
* Re-type copied JSON services to use YLS’s existing JSONDocument and AST types
* Move schema registry/cache/loading behavior directly into YAMLSchemaService
* Update JSON validation to depend on IJSONSchemaService instead of a concrete JSONSchemaService class
